requirements.txt
requests>=2.0.0
python-dotenv>=0.21.0
